Introducing to Machine Learning (
The mobile company Megaline is not satisfied with the fact that many of its customers are still using legacy plans. They want to develop a model that can analyze customer behavior and recommend one of Megaline's new plans: Smart or Ultra.

You have access to subscriber behavior data for customers who have already switched to the new plans (from the Statistical Data Analysis Sprint project). For this classification task, you need to create a model that selects the correct plan. Since the data preprocessing step has already been completed, you can proceed directly to building the model.

Develop a model with the highest possible accuracy. In this project, the accuracy threshold is 0.75. Use the dataset to validate the model's accuracy.